my recipe page
Unfortunately, even after spending 5 days trying to figure out how to work the recipe API at the bottom of the page, I was unable to get it to work. If anyone else in the group finds a way to make it functional, that would be awesome.
P.S.: not sure how it happened, but the max amount of searches was reached for the free plan I got together with the API key and app ID, so you might want to get another one of your own to use in the code (to try and get it to work). Here's the reference website for the recipe API: https://developer.edamam.com/edamam-recipe-api

I have styled the page and added the recipes Jillian sent me the other day.